exract sig parsing to helper function
